Don't pin @cloudflare/workers-types (#258)
This prevents peer dependency conflicts when using templates with C3. C3 ignores the wrangler version specified by the template and always installs the latest version. @cloudflare/workers-types is a peer dependency of wrangler, and the latest version of wrangler may depend on a later version of @cloudflare/workers-types than the version in the template. If @cloudflare/workers-types is pinned to an older version in the template, this causes a peer dependency conflict. Unpinning @cloudflare/workers-types in our templates allows this package to be upgraded to match wrangler, and avoids the conflict in this case.